Additional Base Tests - yiisoft/data/pull/234 - msi 100
Applied  yiisoft/data/pull/234
Moved tests folder into psalm.xml - Level 1 tested
EntityReader's readOne function's limit(1) has been removed to get the msi to 100.

A suitable test needs to be developed to check for the IncrementInteger mutant on the readOne function with the mutant that changes limit(1) to limit(2) still causing a failure with current tests not picking it up mainly due to the \Generator->current returning mixed type where null, array, and object have to be checked in readOne with false also being returned.